+++ crates/common/src/db.rs
@@ -12,87 +12,6 @@ use tracing::{info, warn};
 use crate::config::DatabaseConfig;
 use crate::error::Result;
 
-const V021_MIGRATION_CHECKSUMS: &[(i64, &str, &str)] = &[
-    (
-        20250101000009,
-        "21226d6a5c436c95cfd19277d7f5e4f6f54fc30c9690c28d3f3f4c07343a078b14bdcb0e3f60bc2a2c2b197b716765f2",
-        "ac857a353cc0a325788c54a89ff4dd594af8e13567516a3e334d4d5cc6f38ad8af358001c82ce713367034e79895780b",
-    ),
-    (
-        20250101000013,
-        "275d15eb2f9af232f869eb9a4da30f35a9312f9b8206e9e4ce59a9eb244323a071e39251e31155180a027ceaab3c8788",
-        "3e68a2d74ccc74f7fb182db0ee92b9390e8ca75499fc6b8ac0ea8f2ce629650967d99a7f12f211035cf5ceb80bb63947",
-    ),
-    (
-        20250101000014,
-        "f8d7e71cc4a79bbb69262033a2b881f4110dc53aa39031465c9d2fe1a61c9fc431cec486323baf8265a9bdf8b9d71994",
-        "061a9178e561ca4e5f2c8a814aa326300ba6fe6b1dd40bb6514e1486198a1049aac59529095aa0f864c755463f084f8d",
-    ),
-    (
-        20250101000018,
-        "fa4c9a91965ccd647b6c4372c5db7296d650e7ed4bd717664652174365b85813d417fb2527d84cdebb6f2975c6dbd78d",
-        "429fe44ef8a7d5ae6bcdd1a2f4111e54f660edf17e7c392e15ea69f3ae5236e73a657e291bcdd23547132be083c123cb",
-    ),
-];
-
-async fn bridge_v021_migration_checksums(connection: &mut PgConnection) -> Result<()> {
-    let mut transaction = connection.begin().await?;
-    let history_exists: bool =
-        sqlx::query_scalar("SELECT to_regclass('_sqlx_migrations') IS NOT NULL")
-            .fetch_one(&mut *transaction)
-            .await?;
-
-    if !history_exists {
-        transaction.commit().await?;
-        return Ok(());
-    }
-
-    for &(version, legacy_checksum, current_checksum) in V021_MIGRATION_CHECKSUMS {
-        let checksum_state: Option<(bool, bool, bool)> = sqlx::query_as(
-            r#"
-            SELECT
-                success,
-                checksum = decode($2, 'hex') AS is_legacy,
-                checksum = decode($3, 'hex') AS is_current
-            FROM _sqlx_migrations
-            WHERE version = $1
-            FOR UPDATE
-            "#,
-        )
-        .bind(version)
-        .bind(legacy_checksum)
-        .bind(current_checksum)
-        .fetch_optional(&mut *transaction)
-        .await?;
-
-        let Some((success, is_legacy, is_current)) = checksum_state else {
-            continue;
-        };
-        if !success {
-            return Err(crate::error::Error::invalid_state(format!(
-                "SQLx migration {version} is marked unsuccessful; refusing checksum compatibility rewrite"
-            )));
-        }
-        if !is_legacy && !is_current {
-            return Err(crate::error::Error::invalid_state(format!(
-                "SQLx migration {version} has an unrecognized checksum; refusing compatibility rewrite"
-            )));
-        }
-        if is_legacy {
-            sqlx::query(
-                "UPDATE _sqlx_migrations SET checksum = decode($2, 'hex') WHERE version = $1",
-            )
-            .bind(version)
-            .bind(current_checksum)
-            .execute(&mut *transaction)
-            .await?;
-        }
-    }
-
-    transaction.commit().await?;
-    Ok(())
-}
-
 async fn pin_sqlx_history_search_path(connection: &mut PgConnection) -> Result<()> {
     let (has_attune_history, has_public_history): (bool, bool) = sqlx::query_as(
         r#"
@@ -135,7 +54,6 @@ async fn run_embedded_migrations(connection: &mut PgConnection) -> Result<()> {
 
     let migration_result = async {
         pin_sqlx_history_search_path(connection).await?;
-        bridge_v021_migration_checksums(connection).await?;
         let mut migrator = sqlx::migrate!("../../migrations");
         migrator.set_locking(false);
         migrator.run(&mut *connection).await.map_err(|error| {